+++ samples/bluetooth/peripheral_hids/src/hog.h
